+++ packages/storybook-vue/.storybook/usage-addon/usage.js
@@ -47,6 +47,8 @@ import menuFlyout_en from 'raw-loader!../../stories/3_components/flyout-menu/men
 import menuFlyout_de from 'raw-loader!../../stories/3_components/flyout-menu/menu-flyout_de.md';
 import modal_en from 'raw-loader!../../stories/3_components/modal/modal.md';
 import modal_de from 'raw-loader!../../stories/3_components/modal/modal_de.md';
+import notificationBadge_de from 'raw-loader!../../stories/3_components/notification-badge/notification-badge_de.md';
+import notificationBadge_en from 'raw-loader!../../stories/3_components/notification-badge/notification-badge.md';
 import notificationMessage_de from 'raw-loader!../../stories/3_components/notification-message/notification-message_de.md';
 import notificationMessage_en from 'raw-loader!../../stories/3_components/notification-message/notification-message.md';
 import pagination_en from 'raw-loader!../../stories/3_components/pagination/pagination.md';
@@ -79,8 +81,8 @@ import textField_en from 'raw-loader!../../stories/3_components/text-field/text-
 import textField_de from 'raw-loader!../../stories/3_components/text-field/text-field_de.md';
 import footer_en from 'raw-loader!../../stories/3_components/footer/footer.md';
 import footer_de from 'raw-loader!../../stories/3_components/footer/footer_de.md';
-import badge_de from 'raw-loader!../../stories/3_components/badge/badge_de.md';
-import badge_en from 'raw-loader!../../stories/3_components/badge/badge.md';
+import callout_en from 'raw-loader!../../stories/3_components/callout/callout.md';
+import callout_de from 'raw-loader!../../stories/3_components/callout/callout_de.md';
 import toggleGroup_de from 'raw-loader!../../stories/3_components/toggle-group/toggle-group_de.md';
 import toggleGroup_en from 'raw-loader!../../stories/3_components/toggle-group/toggle-group.md';
 
@@ -121,8 +123,8 @@ const Usage = (props) => {
     'checkbox-group_de': checkboxGroup_de,
     divider_en,
     divider_de,
-    badge_en,
-    badge_de,
+    callout_en,
+    callout_de,
     'data-grid_en': datagrid_en,
     'data-grid_de': datagrid_de,
     'date-picker_en': date_picker_en,
@@ -143,6 +145,8 @@ const Usage = (props) => {
     'flyout-menu_de': menuFlyout_de,
     modal_en,
     modal_de,
+    'notification-badge_en': notificationBadge_en,
+    'notification-badge_de': notificationBadge_de,
     'notification-message_en': notificationMessage_en,
     'notification-message_de': notificationMessage_de,
     pagination_en,